Clarify how to run tests
The existing instructions did not work, because they instructed the
user to start their own Ganache instance, which would then conflict
with the ones started by the scripts.  So clarify this, and also
explain how to run individual tests one by one, which is what the
README already recommended anyway to prevent incorrect nonce errors.