Fix errors in 02_contract.md
I've spotted a few minor errors in the document.

Line 166: if e,d, and c are valid then the weight sums up to 11, not 10.

Line 168: (b) gives a nonce of (1) not (2)

Line 175: There is no vote (c), only (a) and (b). Rename (c) to (b).

I've also unified two styles used to list votes `a.` and `a)` to `(a)` (so it is consistent with the description). Moreover, I have added `- (a)` for a better post-rendering list 
readability.