fix(contracts): save numSignups after merging maciStateAq inside the Poll contract
prevent new signups from changing the record of users being signed for a specific poll. Instead
store this value in the Poll contract after the stateAq is merged. This way, even if a user signs up
after the logs are fetched and the state is reproduced offchain, the proof validation will not fail
because of a numSignups mismatch.

fix #1091